Use float32 as default for polars storage dtypes instead of float64.

+++ eir/data_load/label_setup.py
@@ -522,7 +522,7 @@ def get_file_path_iterator(
 def _get_all_label_columns_and_dtypes(
     cat_columns: Sequence[str],
     con_columns: Sequence[str],
-) -> Tuple[Sequence[str], dict[str, Type[pl.Categorical] | Type[pl.Float64]]]:
+) -> Tuple[Sequence[str], dict[str, Type[pl.Categorical] | Type[pl.Float32]]]:
     supplied_label_columns = _get_column_dtypes(
         cat_columns=cat_columns,
         con_columns=con_columns,
@@ -537,13 +537,13 @@ def _get_all_label_columns_and_dtypes(
 
 def _get_column_dtypes(
     cat_columns: Sequence[str], con_columns: Sequence[str]
-) -> dict[str, Type[pl.Categorical] | Type[pl.Float64]]:
-    dtypes: dict[str, Type[pl.Categorical] | Type[pl.Float64]] = {}
+) -> dict[str, Type[pl.Categorical] | Type[pl.Float32]]:
+    dtypes: dict[str, Type[pl.Categorical] | Type[pl.Float32]] = {}
 
     for cat_column in cat_columns:
         dtypes[cat_column] = pl.Categorical
     for con_column in con_columns:
-        dtypes[con_column] = pl.Float64
+        dtypes[con_column] = pl.Float32
 
     return dtypes
 
@@ -551,7 +551,7 @@ def _get_column_dtypes(
 def _load_label_df(
     label_fpath: Path,
     columns: Sequence[str],
-    dtypes: Optional[dict[str, type[pl.Categorical] | type[pl.Float64]]] = None,
+    dtypes: Optional[dict[str, type[pl.Categorical] | type[pl.Float32]]] = None,
 ) -> pl.DataFrame:
     dtypes = _ensure_id_str_dtype(dtypes=dtypes)
 
@@ -709,7 +709,7 @@ def _check_parsed_label_df(
             continue
 
         assert dtype in [
-            pl.Float64,
+            pl.Float32,
             pl.Categorical,
         ], f"Column {column} has invalid type {dtype}"
 
@@ -876,7 +876,7 @@ def _get_con_manual_vals_dict(
 
     for column in con_columns:
         column_mean = (
-            df.get_column(column).drop_nans().drop_nulls().cast(pl.Float64).mean()
+            df.get_column(column).drop_nans().drop_nulls().cast(pl.Float32).mean()
         )
         if isinstance(column_mean, (int, float)):
             con_means_dict[column] = float(column_mean)
